refactor(controller): remove controller pkg barrel files
