--- packages/controller/src/index.ts
@@ -1,35 +1,0 @@
-// Command dispatcher
-export { CommandDispatcher } from "./core/command-dispatcher.js";
-
-// Core types
-export type { Command, CommandType } from "./core/commands.js";
-export type {
-	ControllerConfig,
-	ControllerMode,
-} from "./core/controller-config.js";
-export { controllerConfigSchema } from "./core/controller-config.js";
-
-// Event bus
-export { EventBus } from "./core/event-bus.js";
-export {
-	deletePidFile,
-	getDaemonPid,
-	isProcessRunning,
-	readPidFile,
-	writePidFile,
-} from "./core/pid-manager.js";
-// State store
-export type {
-	LaunchpadState,
-	SystemState,
-} from "./core/state-store.js";
-export { StateStore } from "./core/state-store.js";
-// Transport system (for future transports like WebSocket, OSC)
-export type { Transport, TransportContext } from "./core/transport.js";
-// IPC client and types for CLI
-export { IPCClient } from "./ipc/ipc-client.js";
-export { LaunchpadController } from "./launchpad-controller.js";
-export type {
-	IPCMessage,
-	IPCResponse,
-} from "./transports/ipc-transport.js";
